assignment/question - 2/myapp/requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
asposestoragecloud==1.0.6
atomicwrites==1.4.0
attrs==19.3.0
awscli==1.18.9
backcall==0.2.0
base58==0.2.2
bleach==3.1.5
boto3==1.14.30
botocore==1.17.30
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
chargebee==2.7.3
click==7.1.2
colorama==0.3.9
ConfigArgParse==1.2.3
coverage==5.1
decorator==4.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-admin-sortable2==0.7.6
django-allauth==0.38.0
django-clear-cache==0.3
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-datetime-utc==1.0.2
django-debug-toolbar==2.2
django-extensions==2.1.4
django-filter==2.1.0
django-filters==0.2.1
django-grappelli==2.14.2
django-guardian==1.4.9
django-haystack==2.8.1
django-postgrespool==0.3.1
django-postgrespool2==1.0.1
django-redis==4.10.0
django-rest-auth==0.9.3
django-rest-framework-mongoengine==3.3.1
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.0
djangorestframework-filters==0.10.2.post0
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.12
entrypoints==0.3
Flask==1.1.2
fpdf==1.7.2
futures==3.0.5
geckodriver-autoinstaller==0.1.0
gevent==20.4.0
geventhttpclient-wheels==1.3.1.dev2
graphviz==0.13.2
greenlet==0.4.15
hjson==1.5.4
idna==2.7
importlib-metadata==1.6.1
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.9.0
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
lambda-packages==0.5.0
langdetect==1.0.8
libmagic==1.0
looptools==1.2.3
lxml==4.4.1
MarkupSafe==1.1.1
mistune==0.8.4
model-bakery==1.1.0
more-itertools==8.4.0
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.16.2
oauthlib==2.1.0
objgraph==3.4.1
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pdfconduit==1.8.10
pdfconduit-convert==1.2.8
pdfconduit-modify==2.2.4
pdfconduit-transform==1.2.2
pdfconduit-utils==1.1.2
pdfrw==0.4
pickleshare==0.7.5
Pillow==7.2.0
PillowImage==1.1.6
pluggy==0.13.1
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
py==1.9.0
pyasn1==0.4.5
PyBundle==1.0.6
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.8.0
PyMuPDF==1.17.4
pypandoc==1.4
pyparsing==2.4.7
PyPDF2==1.26.0
PyPDF3==1.0.1
pyrsistent==0.16.0
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.1
python-decouple==3.1
python-slugify==1.2.0
python3-openid==3.1.0
pytz==2018.7
pywin32==228
pywinpty==0.5.7
PyYAML==5.1
pyzmq==19.0.0
qtconsole==4.7.5
QtPy==1.9.0
redis==3.2.1
reportlab==3.5.46
requests==2.20.0
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.3.3
selenium==3.141.0
Send2Trash==1.5.0
Sendinblue==2.0.5.1
sentry-sdk==0.14.3
sib-api-v3-sdk==6.0.3
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
static3==0.7.0
stripe==2.43.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.4.1
traitlets==4.3.3
twilio==6.44.1
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
xmltodict==0.12.0
zipp==3.1.0
